[PSL-819] Investigate concurrency problems with wallet functions in pasteld, addressing "Insufficient Funds" errors
- added new RPC API to scan wallet for missing transactions:
scanformissingtxs <starting_height>
Scan for missing transactions in the wallet starting from the given height.

Arguments:
1. <starting_height>  (numeric, required)  // starting block height to search for missing txs"

Returns:
[
    "missing_transaction1_txid",
    "missing_transaction2_txid",
    .......
]
