+++ build-aux/vs2022/libbitcoin_wallet/libbitcoin_wallet.vcxproj
@@ -92,6 +92,7 @@
     <ClCompile Include="..\..\..\src\wallet\asyncrpcoperation_shieldcoinbase.cpp" />
     <ClCompile Include="..\..\..\src\wallet\crypter.cpp" />
     <ClCompile Include="..\..\..\src\wallet\db.cpp" />
+    <ClCompile Include="..\..\..\src\wallet\missing_txs.cpp" />
     <ClCompile Include="..\..\..\src\wallet\rpcdump.cpp" />
     <ClCompile Include="..\..\..\src\wallet\rpcwallet.cpp" />
     <ClCompile Include="..\..\..\src\wallet\wallet.cpp" />
@@ -105,6 +106,7 @@
     <ClInclude Include="..\..\..\src\wallet\asyncrpcoperation_shieldcoinbase.h" />
     <ClInclude Include="..\..\..\src\wallet\crypter.h" />
     <ClInclude Include="..\..\..\src\wallet\db.h" />
+    <ClInclude Include="..\..\..\src\wallet\missing_txs.h" />
     <ClInclude Include="..\..\..\src\wallet\rpcwallet.h" />
     <ClInclude Include="..\..\..\src\wallet\wallet.h" />
     <ClInclude Include="..\..\..\src\wallet\walletdb.h" />
