+++ build-aux/vs2022/libbitcoin_wallet/libbitcoin_wallet.vcxproj.filters
@@ -43,6 +43,9 @@
     <ClCompile Include="..\..\..\src\wallet\wallet_ismine.cpp">
       <Filter>Source Files\wallet</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\..\src\wallet\missing_txs.cpp">
+      <Filter>Source Files\wallet</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\..\src\zcbenchmarks.h">
@@ -73,6 +76,9 @@
       <Filter>Source Files\wallet</Filter>
     </ClInclude>
     <ClInclude Include="..\..\..\src\wallet\wallet_ismine.h">
+      <Filter>Source Files\wallet</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\src\wallet\missing_txs.h">
       <Filter>Source Files\wallet</Filter>
     </ClInclude>
   </ItemGroup>
