+++ src/mnode/rpc/tickets-activate.cpp
@@ -1,15 +1,15 @@
-// Copyright (c) 2018-2022 The Pastel Core developers
+// Copyright (c) 2018-2023 The Pastel Core developers
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or https://www.opensource.org/licenses/mit-license.php.
-
 #include <rpc/rpc_consts.h>
 #include <rpc/rpc_parser.h>
 #include <rpc/server.h>
-#include <mnode/rpc/mnode-rpc-utils.h>
+#include <rpc/rpc-utils.h>
 #include <mnode/tickets/nft-act.h>
 #include <mnode/tickets/action-act.h>
 #include <mnode/tickets/collection-act.h>
 #include <mnode/ticket-processor.h>
+#include <mnode/rpc/mnode-rpc-utils.h>
 
 using namespace std;
