+++ src/mnode/rpc/tickets-fake.cpp
@@ -1,12 +1,11 @@
-// Copyright (c) 2018-2022 The Pastel Core developers
+// Copyright (c) 2018-2023 The Pastel Core developers
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or https://www.opensource.org/licenses/mit-license.php.
 #include <rpc/rpc_parser.h>
+#include <rpc/rpc-utils.h>
 #include <mnode/tickets/tickets-all.h>
-#include <mnode/rpc/mnode-rpc-utils.h>
 #include <mnode/rpc/tickets-fake.h>
 #include <mnode/ticket-processor.h>
-#include <main.h>
 
 using namespace std;
