+++ src/mnode/rpc/tickets-list.cpp
@@ -1,13 +1,12 @@
 // Copyright (c) 2018-2023 The Pastel Core developers
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or https://www.opensource.org/licenses/mit-license.php.
-
 #include <enum_util.h>
 #include <rpc/rpc_parser.h>
 #include <rpc/server.h>
+#include <rpc/rpc-utils.h>
 #include <mnode/tickets/tickets-all.h>
 #include <mnode/mnode-controller.h>
-#include <mnode/rpc/mnode-rpc-utils.h>
 #include <mnode/rpc/tickets-list.h>
 
 using namespace std;
