+++ src/mnode/rpc/tickets-register.cpp
@@ -1,9 +1,10 @@
-// Copyright (c) 2018-2022 The Pastel Core developers
+// Copyright (c) 2018-2023 The Pastel Core developers
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or https://www.opensource.org/licenses/mit-license.php.
 #include <rpc/rpc_parser.h>
 #include <rpc/rpc_consts.h>
 #include <rpc/server.h>
+#include <rpc/rpc-utils.h>
 #include <mnode/rpc/mnode-rpc-utils.h>
 #include <mnode/rpc/tickets-activate.h>
 #include <mnode/mnode-controller.h>
